failover is not supported
There was an oversight in commit de737f787e7 that enables failover
option if the version is 17. An existing code checks if there is a
function named pg_create_logical_replication_slot with an argument
"failover". There wasn't such function in Postgres until 17 so it never
adds the "failover" option. Remove this test for Postgres.